# core/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Command-and-control core backend API.",
        "title": "Logos Core API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/channel/sync": {
            "post": {
                "description": "Accepts an inbound.minion_message from a channel module and returns an outbound.minion_message. Payloads are opaque encrypted blobs; core currently returns a no-op outbound payload.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Channels"
                ],
                "summary": "Channel sync (data plane)",
                "parameters": [
                    {
                        "description": "Inbound minion message",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/protocol.InboundMinionMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/protocol.OutboundMinionMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "instance is not a registered module",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "module registry unavailable",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "protocol.InboundMinionMessage": {
            "type": "object",
            "properties": {
                "encrypted_data": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "message_id": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/protocol.MessageMeta"
                },
                "source": {
                    "$ref": "#/definitions/protocol.SourceInfo"
                },
                "timestamp": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "protocol.MessageMeta": {
            "type": "object",
            "additionalProperties": {}
        },
        "protocol.OutboundMinionMessage": {
            "type": "object",
            "properties": {
                "encrypted_data": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "message_id": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/protocol.MessageMeta"
                },
                "timestamp": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "protocol.SourceInfo": {
            "type": "object",
            "properties": {
                "module": {
                    "type": "string"
                },
                "module_instance": {
                    "type": "string"
                },
                "tenant": {
                    "type": "string"
                },
                "transport": {
                    "type": "string"
                }
            }
        },
...
        "responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer JWT token (e.g. \"Bearer eyJhb...\")",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}